+++ src/content-helper/dashboard-widget/top-posts/component-list.tsx
@@ -12,7 +12,6 @@ import DashboardWidgetProvider from '../provider';
 import TopPostListItem from './component-list-item';
 import { TopPostData } from './model';
 import { ContentHelperError } from '../../../blocks/content-helper/content-helper-error';
-import { getDateInUserLang, SHORT_DATE_FORMAT } from '../../../blocks/shared/utils/date';
 
 const FETCH_RETRIES = 3;
 
@@ -23,22 +22,14 @@ function TopPostList() {
 	const [ loading, setLoading ] = useState<boolean>( true );
 	const [ error, setError ] = useState<ContentHelperError>();
 	const [ posts, setPosts ] = useState<TopPostData[]>( [] );
-	const provider = new DashboardWidgetProvider();
 
 	useEffect( () => {
+		const provider = new DashboardWidgetProvider();
+
 		const fetchPosts = async ( retries: number ) => {
 			provider.getTopPosts()
 				.then( ( result ): void => {
-					const mappedPosts: TopPostData[] = result.map(
-						( post: TopPostData ): TopPostData => (
-							{
-								...post,
-								date: getDateInUserLang( new Date( post.date ), SHORT_DATE_FORMAT ),
-							}
-						)
-					);
-
-					setPosts( mappedPosts );
+					setPosts( result );
 					setLoading( false );
 				} )
 				.catch( async ( err ) => {